requirements.txt
click>=8.0
openai>=1.0
pydantic>=2.0
pyyaml>=6.0
requests>=2.28
